requirements.txt
fastapi
uvicorn
playwright
python-dateutil
logging
asyncio
beautifulsoup4
playwright-stealth
lxml
aiohttp
pandas